Corrected put lists

+++ FlavorCart/Controllers/ListsFirestoreController.cs
@@ -63,16 +63,18 @@ public class ListsFirestoreController : ControllerBase
 
     [HttpPut]
     [Route("{id}")]
-    public async Task<ActionResult<User>> UpdateListsAsync(string id)
+    public async Task<ActionResult<User>> UpdateListsAsync(string id, Lists lists)
     {
         try
         {
             var ok = await _usertokenFirestoreController.Verify(Request.Headers["Authorization"].ToString().Remove(0, 7));
             if (ok != null)
             {
-                Lists lists = new Lists();
-                lists.Id = id;
-                
+
+                if (id != lists.Id)
+                {
+                    return BadRequest("Id must match.");
+                }
 
                 return Ok(await _listsRepository.UpdateAsync(lists));
             }
@@ -93,8 +95,10 @@ public class ListsFirestoreController : ControllerBase
             var ok = await _usertokenFirestoreController.Verify(Request.Headers["Authorization"].ToString().Remove(0, 7));
             if (ok != null)
             {
-                Lists lists = new Lists();
-                lists.Id = id;
+                if (id != lists.Id)
+                {
+                    return BadRequest("Id must match.");
+                }
 
                 await _listsRepository.DeleteAsync(lists);
